requirements.txt
fastapi~=0.118.0
passlib~=1.7.4
pydantic~=2.11.9
SQLAlchemy~=2.0.44
python-dotenv~=1.1.1
alembic~=1.17.0
bcrypt==4.3.0
starlette~=0.48.0
uvicorn~=0.30.0
asyncpg>=0.30.0
python-jose~=3.3.0
python-multipart==0.0.20
Pillow>=9.0.0
aiofiles
pydantic-settings>=2.0.0